backend/requirements.txt
fastapi[standard]==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
cryptography==44.0.0
bcrypt==4.2.1
python-multipart==0.0.20
httpx==0.27.0
